refactor(ios): extract workspace runtime picker (MIM-102)
