+++ ios/MimiRemote/Sources/Features/Projects/WorkspaceRootView.swift
@@ -60,53 +60,6 @@ private struct WorkspaceIconMeeGoShape: Shape {
     }
 }
 
-enum WorkspaceSessionRuntimeChoice: String, CaseIterable, Identifiable {
-    case codex
-    case claude
-
-    var id: String { rawValue }
-
-    var runtimeProvider: String {
-        switch self {
-        case .codex:
-            return "codex"
-        case .claude:
-            return "claude"
-        }
-    }
-
-    var listTitle: String {
-        switch self {
-        case .codex:
-            return L10n.text("ui.runtime_default")
-        case .claude:
-            return L10n.text("ui.runtime_claude_short")
-        }
-    }
-
-    var title: String {
-        switch self {
-        case .codex:
-            return L10n.text("ui.create_a_new_codex_session")
-        case .claude:
-            return L10n.text("ui.create_a_new_claude_code_session")
-        }
-    }
-
-    var brandAssetName: String {
-        switch self {
-        case .codex:
-            return "ChatGPT"
-        case .claude:
-            return "Claude"
-        }
-    }
-
-    static func available(claudeChannelAvailable: Bool) -> [Self] {
-        claudeChannelAvailable ? [.codex, .claude] : [.codex]
-    }
-}
-
 enum WorkspaceStripLayout {
     static let horizontalPadding: CGFloat = 24
     // 316pt 能容纳角色头像、路径和一行实时状态，同时在 iPad mini 上露出相邻卡片提示横向滚动。
@@ -1558,75 +1511,6 @@ private struct WorkspaceEmojiPicker: View {
         }
         appearanceStore.setCustomEmoji(emoji, profileID: profileID, projectID: project.id)
         dismiss()
-    }
-}
-
-private struct WorkspaceRuntimePicker: View {
-    @EnvironmentObject private var themeStore: ThemeStore
-    @Environment(\.colorScheme) private var colorScheme
-
-    @Binding var selection: WorkspaceSessionRuntimeChoice
-    let claudeChannelAvailable: Bool
-
-    var body: some View {
-        let tokens = themeStore.tokens(for: colorScheme)
-
-        HStack(spacing: 0) {
-            ForEach(WorkspaceSessionRuntimeChoice.allCases) { choice in
-                let isSelected = selection == choice
-                let isAvailable = choice != .claude || claudeChannelAvailable
-
-                Button {
-                    guard isAvailable else { return }
-                    selection = choice
-                } label: {
-                    HStack(spacing: 6) {
-                        Image(choice.brandAssetName)
-                            .resizable()
-                            .renderingMode(.original)
-                            .scaledToFit()
-                            .frame(width: 17, height: 17)
-                            .accessibilityHidden(true)
-
-                        Text(choice.listTitle)
-                            .font(themeStore.uiFont(.subheadline, weight: isSelected ? .semibold : .medium))
-                            .lineLimit(1)
-                    }
-                    .foregroundStyle(isSelected ? tokens.primaryAction : tokens.secondaryText)
-                    .padding(.horizontal, 12)
-                    .frame(minWidth: 44, minHeight: 36)
-                    .background {
-                        if isSelected {
-                            RoundedRectangle(cornerRadius: 18, style: .continuous)
-                                .fill(tokens.surface)
-                                .shadow(color: Color.black.opacity(0.08), radius: 2, y: 1)
-                        }
-                    }
-                    .contentShape(RoundedRectangle(cornerRadius: 18, style: .continuous))
-                }
-                .buttonStyle(.plain)
-                .disabled(!isAvailable)
-                .opacity(isAvailable ? 1 : 0.42)
-                .accessibilityLabel(choice.listTitle)
-                .accessibilityAddTraits(isSelected ? .isSelected : [])
-                .accessibilityHint(
-                    isAvailable
-                        ? L10n.text("ui.show_runtime_sessions_hint")
-                        : L10n.text("ui.runtime_unavailable_hint")
-                )
-            }
-        }
-        .padding(4)
-        .frame(minHeight: 44)
-        .background(tokens.elevatedSurface, in: RoundedRectangle(cornerRadius: 22, style: .continuous))
-        .overlay {
-            RoundedRectangle(cornerRadius: 22, style: .continuous)
-                .stroke(tokens.border.opacity(0.72), lineWidth: 1)
-        }
-        .fixedSize(horizontal: true, vertical: false)
-        .accessibilityElement(children: .contain)
-        .accessibilityLabel(L10n.text("ui.runtime_provider"))
-        .accessibilityIdentifier("workspace.sessions.runtimePicker")
     }
 }
